+++ circonus/resource_circonus_graph.go
@@ -9,7 +9,7 @@ import (
 	api "github.com/circonus-labs/go-apiclient"
 	"github.com/circonus-labs/go-apiclient/config"
 	"github.com/hashicorp/errwrap"
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
 const (
@@ -481,15 +481,15 @@ func graphRead(d *schema.ResourceData, meta interface{}) error {
 		rightAxisMap[string(graphAxisMinAttr)] = strconv.FormatFloat(*g.MinRightY, 'f', -1, 64)
 	}
 
-	d.Set(graphDescriptionAttr, g.Description)
+	_ = d.Set(graphDescriptionAttr, g.Description)
 
 	if err := d.Set(graphLeftAttr, leftAxisMap); err != nil {
 		return errwrap.Wrapf(fmt.Sprintf("Unable to store graph %q attribute: {{err}}", graphLeftAttr), err)
 	}
 
-	d.Set(graphLineStyleAttr, g.LineStyle)
-	d.Set(graphNameAttr, g.Title)
-	d.Set(graphNotesAttr, indirect(g.Notes))
+	_ = d.Set(graphLineStyleAttr, g.LineStyle)
+	_ = d.Set(graphNameAttr, g.Title)
+	_ = d.Set(graphNotesAttr, indirect(g.Notes))
 
 	if err := d.Set(graphRightAttr, rightAxisMap); err != nil {
 		return errwrap.Wrapf(fmt.Sprintf("Unable to store graph %q attribute: {{err}}", graphRightAttr), err)
@@ -503,7 +503,7 @@ func graphRead(d *schema.ResourceData, meta interface{}) error {
 		return errwrap.Wrapf(fmt.Sprintf("Unable to store graph %q attribute: {{err}}", graphMetricClusterAttr), err)
 	}
 
-	d.Set(graphStyleAttr, g.Style)
+	_ = d.Set(graphStyleAttr, g.Style)
 
 	if err := d.Set(graphTagsAttr, tagsToState(apiToTags(g.Tags))); err != nil {
 		return errwrap.Wrapf(fmt.Sprintf("Unable to store graph %q attribute: {{err}}", graphTagsAttr), err)
@@ -621,12 +621,12 @@ func (g *circonusGraph) ParseConfig(d *schema.ResourceData) error {
 	}
 
 	if v, found := d.GetOk(graphLineStyleAttr); found {
-		switch v.(type) {
+		switch v := v.(type) {
 		case string:
-			s := v.(string)
+			s := v
 			g.LineStyle = &s
 		case *string:
-			g.LineStyle = v.(*string)
+			g.LineStyle = v
 		default:
 			return fmt.Errorf("PROVIDER BUG: unsupported type for %q: %T", graphLineStyleAttr, v)
 		}
@@ -685,12 +685,12 @@ func (g *circonusGraph) ParseConfig(d *schema.ResourceData) error {
 			}
 
 			if v, found := metricAttrs[graphMetricFormulaAttr]; found {
-				switch v.(type) {
+				switch v := v.(type) {
 				case string:
-					s := v.(string)
+					s := v
 					datapoint.DataFormula = &s
 				case *string:
-					datapoint.DataFormula = v.(*string)
+					datapoint.DataFormula = v
 				default:
 					return fmt.Errorf("PROVIDER BUG: unsupported type for %q: %T", graphMetricAttr, v)
 				}
@@ -809,24 +809,24 @@ func (g *circonusGraph) ParseConfig(d *schema.ResourceData) error {
 			}
 
 			if v, found := metricClusterAttrs[graphMetricFormulaAttr]; found {
-				switch v.(type) {
+				switch v := v.(type) {
 				case string:
-					s := v.(string)
+					s := v
 					metricCluster.DataFormula = &s
 				case *string:
-					metricCluster.DataFormula = v.(*string)
+					metricCluster.DataFormula = v
 				default:
 					return fmt.Errorf("PROVIDER BUG: unsupported type for %q: %T", graphMetricFormulaAttr, v)
 				}
 			}
 
 			if v, found := metricClusterAttrs[graphMetricFormulaLegendAttr]; found {
-				switch v.(type) {
+				switch v := v.(type) {
 				case string:
-					s := v.(string)
+					s := v
 					metricCluster.LegendFormula = &s
 				case *string:
-					metricCluster.LegendFormula = v.(*string)
+					metricCluster.LegendFormula = v
 				default:
 					return fmt.Errorf("PROVIDER BUG: unsupported type for %q: %T", graphMetricFormulaLegendAttr, v)
 				}
@@ -871,12 +871,12 @@ func (g *circonusGraph) ParseConfig(d *schema.ResourceData) error {
 	}
 
 	if v, found := d.GetOk(graphStyleAttr); found {
-		switch v.(type) {
+		switch v := v.(type) {
 		case string:
-			s := v.(string)
+			s := v
 			g.Style = &s
 		case *string:
-			g.Style = v.(*string)
+			g.Style = v
 		default:
 			return fmt.Errorf("PROVIDER BUG: unsupported type for %q: %T", graphStyleAttr, v)
 		}